+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Changelog
 
+## [1.16.1-2.1.1] - 2020-08-06
+- Fixed an issue where mobs would get stuck in a block for a while.
+- Mobs will now avoid fire blocks.
+
 ## [1.16.1-2.1.0] - 2020-07-15
-- Added new Nether Blocks (8)
+- Added new Nether Blocks (10)
 - Tweaked all wall torches model (texture mappings)
 - Support for MOD update notifications
 - Created change history log